fix a bug for generating VCF compatible REF/ALT
