+++ pgr-db/src/aln.rs
@@ -5,6 +5,7 @@ use log::debug;
 use rustc_hash::{FxHashMap, FxHashSet};
 use std::collections::HashSet;
 use wavefront_aln::*;
+use  std::cmp::Ordering;
 
 pub type HitPair = ((u32, u32, u8), (u32, u32, u8)); //(bgn1, end1, orientation1),  (bgn2, end2, orientation2)
 
@@ -305,16 +306,26 @@ pub fn get_variants_from_aln_pair_map(
     query_str: &str,
 ) -> Vec<(u32, u32, char, String, String)> {
     let mut current_variant = Vec::<(char, char, char)>::new();
+
     let aggregate_variants = |previous_match: &(u32, u32, char, char, char),
                               current_variant: &Vec<(char, char, char)>|
      -> Option<(u32, u32, char, String, String)> {
         let t_variant_segment = String::from_iter(current_variant.iter().map(|v| v.0));
         let q_variant_segment = String::from_iter(current_variant.iter().map(|v| v.1));
-        let v_type = current_variant[0].2;
+        let t_variant_segment = t_variant_segment.replace('-', "").trim().to_string();
+        let q_variant_segment = q_variant_segment.replace('-', "").trim().to_string();
+        let t_len = t_variant_segment.len();
+        let q_len = q_variant_segment.len();
+        let v_type = match t_len.cmp(&q_len) {
+            Ordering::Greater => 'I',
+            Ordering::Less => 'D',
+            Ordering::Equal => 'X',
+        };
+
         match v_type {
             'X' => Some((
-                previous_match.0 + 1,
-                previous_match.1 + 1,
+                previous_match.0+1,
+                previous_match.1+1,
                 'X',
                 t_variant_segment,
                 q_variant_segment,
@@ -345,55 +356,27 @@ pub fn get_variants_from_aln_pair_map(
             let q_char = query_str.as_bytes()[q_pos as usize] as char;
             if !current_variant.is_empty() {
                 variants.push(aggregate_variants(&previous_match, &current_variant));
-                current_variant.clear();
             };
-            previous_match = (t_pos, q_pos, 'M', t_char, q_char);
+
+            current_variant.clear();
             debug!("{} {} {:1} {:1} {}", t_pos, q_pos, t_char, q_char, t);
+            previous_match = (t_pos, q_pos, 'M', t_char, q_char);
         }
         'X' => {
             let t_char = target_str.as_bytes()[t_pos as usize] as char;
             let q_char = query_str.as_bytes()[q_pos as usize] as char;
-            if !current_variant.is_empty() {
-                if current_variant.last().unwrap().2 == t {
-                    current_variant.push((t_char, q_char, t));
-                } else {
-                    variants.push(aggregate_variants(&previous_match, &current_variant));
-                    current_variant.clear();
-                    current_variant.push((t_char, q_char, t));
-                }
-            } else {
-                current_variant.push((t_char, q_char, t));
-            }
+            debug!("{} {} {:1} {:1} {}", t_pos, q_pos, t_char, q_char, t);
+            current_variant.push((t_char, q_char, t));
         }
         'I' => {
             let q_char = query_str.as_bytes()[q_pos as usize] as char;
             debug!("{} {} {:1} {:1} {}", t_pos, q_pos, '-', q_char, t);
-            if !current_variant.is_empty() {
-                if current_variant.last().unwrap().2 == t {
-                    current_variant.push(('-', q_char, t));
-                } else {
-                    variants.push(aggregate_variants(&previous_match, &current_variant));
-                    current_variant.clear();
-                    current_variant.push(('-', q_char, t));
-                }
-            } else {
-                current_variant.push(('-', q_char, t));
-            }
+            current_variant.push(('-', q_char, t));
         }
         'D' => {
             let t_char = target_str.as_bytes()[t_pos as usize] as char;
             debug!("{} {} {:1} {:1} {}", t_pos, q_pos, t_char, '-', t);
-            if !current_variant.is_empty() {
-                if current_variant.last().unwrap().2 == t {
-                    current_variant.push((t_char, '-', t));
-                } else {
-                    variants.push(aggregate_variants(&previous_match, &current_variant));
-                    current_variant.clear();
-                    current_variant.push((t_char, '-', t));
-                }
-            } else {
-                current_variant.push((t_char, '-', t));
-            }
+            current_variant.push((t_char, '-', t));
         }
         _ => {}
     });
@@ -401,6 +384,72 @@ pub fn get_variants_from_aln_pair_map(
         variants.push(aggregate_variants(&previous_match, &current_variant));
     };
     variants.into_iter().flatten().collect::<Vec<_>>()
+}
+
+type AlignmentResult = Vec<(u32, u32, char, String, String)>;
+pub fn get_variant_segments(
+    target_str: &[u8],
+    query_str: &[u8],
+    left_padding: usize,
+    max_wf_length: Option<u32>,
+    mismatch_penalty: i32,
+    open_penalty: i32,
+    extension_penalty: i32,
+) -> Option<AlignmentResult> {
+    let set_len_diff = (query_str.len() as i64 - target_str.len() as i64).unsigned_abs() as u32;
+    let max_wf_length = if let Some(max_wf_length) = max_wf_length {
+        max_wf_length
+    } else {
+        std::cmp::max(2 * set_len_diff, 128_u32)
+    };
+
+    // we need to reverse the string for alignment such the the gaps are on the left
+    // maybe we can do this in the stack for performance in the future
+    // we assume the left_padding base on the left side are identical
+    let mut r_t_str = target_str[left_padding..].to_vec();
+    let mut r_q_str = query_str[left_padding..].to_vec();
+    r_t_str.reverse();
+    r_q_str.reverse();
+    let r_t_str = String::from_utf8_lossy(&r_t_str[..]);
+    let r_q_str = String::from_utf8_lossy(&r_q_str[..]);
+    let t_len_minus_one = left_padding as u32 + r_t_str.len() as u32 - 1;
+    let q_len_minus_one = left_padding as u32 + r_q_str.len() as u32 - 1;
+
+    if let Some((aln_target_str, aln_query_str)) = wfa_align_bases(
+        &r_t_str,
+        &r_q_str,
+        max_wf_length,
+        mismatch_penalty,
+        open_penalty,
+        extension_penalty,
+    ) {
+        let mut aln_pairs = wfa_aln_pair_map(&aln_target_str, &aln_query_str);
+        // assume the base on the left are identical  ( # of base = left_padding)
+        (0..left_padding).for_each(|delta| {
+            aln_pairs.push((
+                (r_t_str.len() + delta) as u32,
+                (r_q_str.len() + delta) as u32,
+                'M',
+            ));
+        });
+        // convert the coordinate from the reverse to the forward sequence
+        aln_pairs.iter_mut().for_each(|(t_pos, q_pos, _c)| {
+            *t_pos = t_len_minus_one - *t_pos;
+            *q_pos = q_len_minus_one - *q_pos;
+        });
+        aln_pairs.reverse();
+
+        // compute the VCF like variant representation
+        let target_str = String::from_utf8_lossy(target_str);
+        let query_str = String::from_utf8_lossy(query_str);
+        Some(get_variants_from_aln_pair_map(
+            &aln_pairs,
+            &target_str,
+            &query_str,
+        ))
+    } else {
+        None
+    }
 }
 
 #[cfg(test)]
@@ -451,5 +500,27 @@ mod test {
                 println!("{} {} {} {} {}", t_pos, q_pos, t, s1, s2);
             });
         }
+    }
+
+    #[test]
+    fn test_aggreate_variant() {
+        use crate::aln::{
+            get_variant_segments, get_variants_from_aln_pair_map, wfa_align_bases, wfa_aln_pair_map,
+        };
+        use log::debug;
+        use simple_logger::SimpleLogger;
+        SimpleLogger::new().init().unwrap();
+        let t_str =
+            "ACGGAGGTGAGCCTGGGAGCATAGAGGTGGGCCTGGGAGCATGGCGGCGGGGGGGGGGCCTGGGAGCACAGGGCGGGCC";
+        let q_str =
+            "ACGGAGGTGAGCCTGGGAGCATAGAGGTGGGCCTGGGAGCATGGCGGTGGGGGGGGGCCTGGGAGCACAGGGCGGGCC";
+
+        if let Some(aln_res) =
+            get_variant_segments(t_str.as_bytes(), q_str.as_bytes(), 1, Some(128), 3, 3, 1)
+        {
+            aln_res
+                .into_iter()
+                .for_each(|v| println!("{} {} {} {} {}", v.0, v.1, v.2, v.3, v.4));
+        };
     }
 }
